# static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ПайсМобайл — фактическая API-документация",
    "version": "1.0.0",
    "description": "Документация составлена по фактическому коду Flask-приложения. Большинство маршрутов возвращает HTML или HTTP 302, не REST JSON. JSON-ответы есть у PUT /select-city/{city}, PUT /profile, PUT /save-card, DELETE /delete-card и DELETE /cancel-tariff."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5001",
      "description": "Локальный запуск Flask-приложения"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Стандартная подписанная Flask session-cookie. Получается после успешного POST /login или POST /register."
      }
    },
    "schemas": {
      "JsonResponse": {
        "type": "object",
        "required": [
          "ok",
          "message"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Настройки профиля сохранены."
          },
          "city": {
            "type": "string",
            "example": "Москва",
            "description": "Возвращается в /select-city/{city}."
          },
          "last4": {
            "type": "string",
            "example": "1111",
            "description": "Возвращается при сохранении карты."
          }
        }
      },
      "LoginForm": {
        "type": "object",
        "required": [
          "login",
          "password"
        ],
        "properties": {
          "login": {
            "type": "string",
            "description": "Телефон или email пользователя.",
            "example": "alex@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "Password!1"
          }
        }
      },
...
      "PromoCodeForm": {
        "type": "object",
        "properties": {
          "promo": {
            "type": "string",
            "description": "Любой непустой промокод считается принятым. Пустое значение даёт warning.",
            "example": "SALE2026"
          },
          "promo_id": {
            "type": "string",
            "description": "ID акции. Если указан, приоритетнее поля promo.",
            "example": "3"
          },
          "friend_phone": {
            "type": "string",
            "description": "Используется акцией cashback.",
            "example": "+79990001122"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/tariffs": {
      "get": {
        "tags": [
          "Города и тарифы"
        ],
        "summary": "Страница тарифов и конструктор тарифа",
        "description": "Возвращает HTML со всеми тарифами, выбранным городом, доступностью тарифов по городу и расчётом цены конструктора. При наличии query-параметра city он сохраняется в сессию selected_city.",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "Москва"
            },
            "description": "Выбранный город. Если не передан: берётся город пользователя, затем selected_city из сессии, затем Москва."
          },
          {
            "name": "minutes",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 500,
              "example": 500
            },
            "description": "Минуты в конструкторе. Невалидное значение заменяется на 500."
          },
          {
            "name": "internet",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "example": 20
            },
            "description": "ГБ в конструкторе. Невалидное значение заменяется на 20."
          },
          {
            "name": "sms",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 100,
              "example": 100
            },
            "description": "SMS в конструкторе. Невалидное значение заменяется на 100."
          }
        ],
        "responses": {
          "200": {
            "description": "HTML-страница успешно сформирована.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                },
                "example": "<!doctype html><html lang=\"ru\">...</html>"
              }
            }
          }
        }
      }
    },
    "/cities": {
      "get": {
        "tags": [
          "Города и тарифы"
        ],
        "summary": "Страница выбора города",
        "description": "Возвращает HTML со справочником городов, сгруппированных по регионам. Поиск q выполняется как подстрока по названию города и области в нижнем регистре.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "мос"
            },
            "description": "Строка поиска."
          }
        ],
        "responses": {
          "200": {
            "description": "HTML-страница успешно сформирована.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                },
                "example": "<!doctype html><html lang=\"ru\">...</html>"
              }
            }
          }
        }
      }
    },
    "/select-city/{city}": {
      "get": {
        "tags": [
          "Города и тарифы"
        ],
        "summary": "Выбор города через ссылку",
        "description": "Сохраняет city в сессию selected_city. Если пользователь авторизован, обновляет город в профиле без проверки, что город есть в справочнике. Возвращает redirect на Referer или /cities.",
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Москва"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Перенаправление браузера. Фактическая цель зависит от сценария: next, profile, my-tariff, login, tariffs и т.д.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Города и тарифы"
        ],
        "summary": "Выбор города с JSON-ответом",
        "description": "То же действие, что GET, но возвращает JSON. Авторизация не обязательна.",
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Москва"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Город сохранён в сессии.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonResponse"
                },
                "example": {
                  "ok": true,
                  "message": "Выбран город: Москва.",
                  "city": "Москва"
                }
              }
            }
          }
        }
      }
    },
    "/actions": {
      "get": {
        "tags": [
          "Акции"
        ],
        "summary": "Страница акций",
        "description": "Возвращает HTML со списком акций из data/promos.json. Кнопки истёкших акций отключены только на уровне интерфейса; backend отдельно expired не проверяет.",
        "responses": {
          "200": {
            "description": "HTML-страница успешно сформирована.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                },
                "example": "<!doctype html><html lang=\"ru\">...</html>"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Акции"
        ],
        "summary": "Активация акции или промокода",
        "description": "Если передан promo_id, приложение пытается активировать акцию и требует авторизацию. Если promo_id пустой, но promo непустой — любой промокод принимается. Ответ всегда через flash + redirect, JSON не возвращается.",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/PromoCodeForm"
              }
            }
          }
        },
        "responses": {
          "302": {
            "description": "Перенаправление браузера. Фактическая цель зависит от сценария: next, profile, my-tariff, login, tariffs и т.д.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "get": {
        "tags": [
          "Авторизация"
        ],
        "summary": "Форма входа",
        "description": "Возвращает HTML-форму входа.",
        "responses": {
          "200": {
            "description": "HTML-страница успешно сформирована.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                },
                "example": "<!doctype html><html lang=\"ru\">...</html>"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Авторизация"
        ],
        "summary": "Вход в личный кабинет",
        "description": "Ищет пользователя по email или телефону, проверяет пароль и записывает user_id в Flask-сессию. При успехе redirect на query next или /my-tariff. При ошибке возвращает HTML формы входа со статусом 200.",
        "parameters": [
          {
            "name": "next",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "/profile"
            },
            "description": "Куда перенаправить после успешного входа."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/LoginForm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "HTML-страница успешно сформирована.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                },
                "example": "<!doctype html><html lang=\"ru\">...</html>"
              }
            }
          },
          "302": {
            "description": "Перенаправление браузера. Фактическая цель зависит от сценария: next, profile, my-tariff, login, tariffs и т.д.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/register": {
      "get": {
        "tags": [
          "Авторизация"
        ],
        "summary": "Форма регистрации",
        "description": "Возвращает HTML-форму регистрации и список городов.",
        "responses": {
          "200": {
            "description": "HTML-страница успешно сформирована.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                },
                "example": "<!doctype html><html lang=\"ru\">...</html>"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Авторизация"
        ],
        "summary": "Регистрация пользователя",
        "description": "Создаёт пользователя в data/users.json. При успехе записывает user_id в сессию и перенаправляет на query next или /profile. При ошибке возвращает HTML формы со статусом 200.",
        "parameters": [
          {
            "name": "next",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "/payment/serafim"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/RegisterForm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "HTML-страница успешно сформирована.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                },
                "example": "<!doctype html><html lang=\"ru\">...</html>"
              }
            }
          },
          "302": {
            "description": "Перенаправление браузера. Фактическая цель зависит от сценария: next, profile, my-tariff, login, tariffs и т.д.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/my-tariff": {
      "get": {
        "tags": [
          "Кабинет"
        ],
        "summary": "Мой тариф",
        "description": "Возвращает HTML текущего тарифа, баланс, extra_gb, последнюю оплату тарифа и дату следующего списания. Если тарифа нет — показывает состояние без подключения.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "HTML-страница успешно сформирована.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                },
                "example": "<!doctype html><html lang=\"ru\">...</html>"
              }
            }
          },
          "302": {
            "description": "Пользователь не авторизован. Flask добавляет flash-сообщение и перенаправляет на /login?next=<исходный_путь>.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string",
                  "example": "/login?next=/profile"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}